Display on the page a list of programmes
Functionality to display a list of programmes on the page, showing
title, image and small synopses.
+ Unit tests

Skipped tests that deal with the document, as right now I am running the
tests from the command line. Investigating how to make running tests on
the browser work! (Should have thought of that earlier...)