Add A to Z navigation
Pagination of a letter is done through buttons and not added to the
querystring, but for the letters I have decided to direct it through the
url. Better for going back to whatever letter you were looking at
before. (For the pagination inside a letter decided on buttons so the
page doesn't reload every time. Can't save the page you were on this
way, but I think it looks nicer).
Obviously, both paginations could done via links or via buttons. Depends
on what the user prefers.